YOURBOT/requirements.txt
customtkinter==5.2.0
openai==0.27.8
Pillow==10.0.0
